Rename client node module and Node class.
