+++ tiled/client/__init__.py
@@ -1,4 +1,4 @@
 from .constructors import from_context, from_profile, from_uri  # noqa: F401
+from .container import ASCENDING, DESCENDING  # noqa: F401
 from .context import Context  # noqa: F401
-from .node import ASCENDING, DESCENDING  # noqa: F401
 from .utils import hide_logs, record_history, show_logs  # noqa: F401
